# src/main/resources/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/review/register-review": {
      "post": {
        "tags": [
          "review-controller"
        ],
        "summary": "Register new review",
        "operationId": "createReview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Review registered in the system",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRegistrationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Review registration failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRegistrationResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRegistrationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/review/edit-review/{id}": {
      "patch": {
        "tags": [
          "review-controller"
        ],
        "summary": "Edit review",
        "operationId": "updateReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Edit review in the system",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "Edit review failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/review/search": {
      "get": {
        "tags": [
          "review-controller"
        ],
        "summary": "Search reviews by book title",
        "description": "Returns a list of reviews for books matching the given title (partial match)",
        "operationId": "searchReviewsByBookTitle",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "description": "DDD",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Clean Code"
          }
        ],
        "responses": {
          "200": {
            "description": "List of reviews found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No reviews found for the given book title",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameter",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/review/list-only-review/{id}": {
      "get": {
        "tags": [
          "review-controller"
        ],
        "summary": "List reviews for id",
        "operationId": "listReviewForId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "LList review in the system for id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "List review failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/review/list-all-reviews": {
      "get": {
        "tags": [
          "review-controller"
        ],
        "summary": "List all reviews",
        "operationId": "listAllReviews",
        "responses": {
          "200": {
            "description": "List all review in the system",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "List all review failed",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/review/delete-review/{id}": {
      "delete": {
        "tags": [
          "review-controller"
        ],
        "summary": "Delete review",
        "operationId": "deleteReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Delete deleted successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRegistrationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Delete review failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRegistrationResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRegistrationResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ReviewRequest": {
        "type": "object",
        "properties": {
          "reviewTitle": {
            "type": "string"
          },
          "bookName": {
            "type": "string"
          },
          "review": {
            "type": "string"
          },
          "bookNote": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ReviewRegistrationResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ReviewResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "reviewTitle": {
            "type": "string"
          },
          "bookName": {
            "type": "string"
          },
          "review": {
            "type": "string"
          },
          "bookNote": {
            "type": "integer",
            "format": "int64"
          }
        }
      }
    }
  }
}